# Conflicts:
#	sdk/src/main/scala/com/horizen/account/forger/AccountForgeMessageBuilder.scala

+++ sdk/src/main/scala/com/horizen/account/block/AccountBlock.scala
@@ -11,13 +11,13 @@ import com.horizen.evm.TrieHasher
 import com.horizen.proof.{Signature25519, VrfProof}
 import com.horizen.secret.PrivateKey25519
 import com.horizen.serialization.Views
-import com.horizen.utils.{ByteArrayWrapper, BytesUtils, MerklePath}
+import com.horizen.utils.{BytesUtils, MerklePath}
 import com.horizen.validation.InconsistentSidechainBlockDataException
 import com.horizen.{ScorexEncoding, SidechainTypes, account}
 import scorex.core.block.Block
-import scorex.util.{ScorexLogging, idToBytes}
-import com.horizen.account.utils.Account
+import scorex.util.{ScorexLogging}
 
+import java.math.BigInteger
 import scala.util.Try
 
 @JsonView(Array(classOf[Views.Default]))
@@ -90,7 +90,7 @@ object AccountBlock extends ScorexEncoding {
              stateRoot: Array[Byte],
              receiptsRoot: Array[Byte],
              forgerAddress: AddressProposition,
-             baseFee: Long,
+             baseFee: BigInteger,
              gasUsed: Long,
              gasLimit: Long,
              companion: SidechainAccountTransactionsCompanion,
